refactor: create the `frame_manager` crate
To test the frame manager. It is not possible to do tests in the
`kernel` crate because custom targets do not have their own `std`
crates.
