+++ frame_manager/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "frame_manager"
+version = "0.1.0"
+authors = ["toku-sa-n <tokusan441@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bit_field = "0.10.1"
+os_units = "0.2.7"
+uefi = "0.8.0"
+x86_64 = "0.13.3"
